requirements.txt
requests
# MCP Python SDK
mcp
# For HTTP/SSE transport
uvicorn
starlette
# HTTP client
httpx
#.env variables
python-dotenv
# Testing
pytest

